Foreign key type should match the referencing field

+++ database/migrations/2021_04_13_135939_create_admissions_tables_table.php
@@ -15,8 +15,8 @@ class CreateAdmissionsTablesTable extends Migration
     {
         Schema::create('admissions_tables', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('dossier_id')->nullable();
-            $table->unsignedInteger('country_id')->nullable();
+            $table->unsignedBigInteger('dossier_id')->nullable();
+            $table->unsignedBigInteger('country_id')->nullable();
             $table->integer('year')->nullable();
             $table->timestamps();
         });
